Create cpu_pipeline.v
Well, it's 4am and here's where I'm at.

The single cycle worked for my ALU_test.v, matched up everything in the registers. Still not good on the branch and whatnot, need to implement this.

I started on the pipeline and it seems (sorta) straight forward, just need to connect all the block outputs to the correct module (ID, EX. MEM, etc). Also need to finish passing along the controls through the blocks, I've added some detail about this in the file.

According to the grading, it seems that the pipeline isn't necessarily a requirement? it will only improve our CPI... So I would focus on getting the single cycle working and then if we have time get the pipeline working.

I have an exam on Monday so I'll be spending all my time tomorrow studying for that. I'll be studying at the CAE all day so if you want, I can study while you work on this and we can still bounce ideas back and forth. I can also show you how I got all my testing set up, but I've been using modelsim so we'll see.

I don't know why I'm typing this here. I'm to lazy to type and email. I'll just talk to you tomorrow.